requirements.txt
Cython==3.0.10
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
isodate==0.6.1
lxml==5.1.1
multidict==6.0.5
numpy==1.24.4
pandas==1.5.3
platformdirs==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
six==1.16.0
tabulate==0.9.0
urllib3==2.2.1
yarl==1.9.4
zeep==4.2.1
aiohttp==3.9.3
noaa-coops==0.3.2
